Implement basic Nix Flake for integration tests
We're going to implement a small API proxy which makes sure that we
don't hand out the Trello API token and secret to people having access
to the exported game.

Right now, the Flake is pretty rudimentary, but it does integrate a
project called "Gut", which we're going to use for triggering the API
request.

Implementing this has been a huge pain for mainly three reasons:

  * Running Godot in a headless fashion isn't so easy, because while
    there is a headless build platform, it also diverges from platforms
    with a UI in some ways, eg. OS.get_unique_id() returning an empty
    string. On the other hand, using Xvfb doesn't work either, since
    Godot doesn't work with software rasterisation.
  * Godot assumes that there is only one path for resources and there is
    no way to eg. specify a search path as it's common with most
    language ecosystems.
  * Certificates are bundled with Godot and getting around that only
    works by setting network/ssl/certificates in project settings. The
    upstreom issue is https://github.com/godotengine/godot/issues/22232.

We're addressing the second issue, by selectively remapping the paths to
Gut in Godot itself. This (and the last issue) unfortunately also means
that we need to patch Godot, since remapping paths only works in a
global way.

The first issue is "solved" by simply going the headless route and fix
our code to work with it, because the headless (and thus "server")
platform is more or less a minimal environment which is useful to ensure
that the code really works on all platforms.

For the moment, I just added a small stub test implementation, which
just performs the necessary UI actions and waits until a message is
displayed, indicating succees.

Unfortunately, there is no signal for whenever the text of a Label node
changes, so we're using a small for-loop with a timeout of 60 seconds to
yield until we get the expected message.

Signed-off-by: aszlig <aszlig@nix.build>