Add very basic Trello API simulator
This is a small Flask application which tries to mimic only the API
endpoints that we need from Trello.

The idea here is that we use this simulator to implement a test that
verifies whether the current implementation works and when replacing
this with our intermediate proxy, we can make sure that API calls are
still behaving the same way.

Our simulator also has one extra endpoint called /__get_state which can
be used for our test implementation to verify whether the remote state
matches the values we have submitted.

Signed-off-by: aszlig <aszlig@nix.build>