Implement full NixOS VM integration test
This essentially consists of network of 3 virtual machines:

  * A "Trello" server running nginx with the "api.trello.com" domain
    serving our Trello simulator.
  * An Apache node, which currently is doing nothing but is going to be
    running our soon-to-be-written intermediate proxy.
  * The client running the Trello reporting tool.
  * A DNS server which provides a root zone for all the domains.

All of the nginx nodes have snakeoil certificates, so that we can
properly use TLS.

Our test script now also overrides the variables for the token, key and
list_id since the values defined in the actual sorce code contain
invalid values. The latter is deliberate (I think) since the user of the
tool is supposed to change these values.

Additionally, after the UI indicates success, we now also use the
special /__get_state route to fetch the remote state and check whether
the data we submitted was actually properly processed by our Trello
simulator.

With this done, we con now finally start the implementation of our
intermediate proxy and have a test to verify the correctness of the
implementation.

Note that the reason why we're using Apache instead of nginx for our
proxy node is to make sure we more closely resemble the environment of a
shared webhosting provider.

Signed-off-by: aszlig <aszlig@nix.build>