Add small Trello API proxy in PHP
The main reason why I'm using PHP here is because that's what you'll
find on most shared webhosting environment. In order to properly
generate the multipart/form-data content-type, I'm also relying on the
cURL extension, which fortunately is also pretty common in these
environments.

While the proxy is already integrated with our Apache test node, we
currently don't use it yet, because we first need to change the
reporting tool to use it.

Signed-off-by: aszlig <aszlig@nix.build>