Integrate PHP proxy into reporting tool
This is essentially a small refactor of the reporting tool, where we now
just send a single HTTP request to create the Trello card, add a label
and one or more attachments.

I also added a little bit of error handling, so that for example if the
HTTP connection fails or timeouts, the user actually gets feedback about
that instead of getting no feedback at all.

Additionally, I renamed the proxy.test domain to proxy.example, since
this means that we no longer need to change the example domain name and
can directly use it in testing, even though the IANA recommends[1] to use
the ".test" top-level domain.

[1]: https://tools.ietf.org/html/rfc2606#page-2

Signed-off-by: aszlig <aszlig@nix.build>