Fix API implementation against real Trello API
It turns out that the upstream Trello API documentation[1] is either
plain wrong or heavily outdated. For example most of the endpoints do
return the objects they were working on and a lot of the identifiers do
not have a length of 32 bytes but 24.

Additionally, a few types were wrong, for example the "bytes" field in
attachments is actually the size of the file in bytes. Since the API
documentation lists it as type "string" I assumed it just contained the
raw bytes of the data.

We were using the bytes field to store a checksum of the uploaded file,
so we can verify it in the test. With the type being an integer we no
longer can do that, so I added another "chksum" field, which is only
returned in /__get_state.

[1]: https://developer.atlassian.com/cloud/trello/rest/

Signed-off-by: aszlig <aszlig@nix.build>