Handle and show proxy error in the UI
So far we just assumed that everything went fine whenever we
successfully sent request, however we didn't actually check the HTTP
status code, so the "Feedback sent successfully" message would end up
being a big lie.

To address this, we new not only check whether the HTTP status code is
HTTP 200, but we also fetch the body and display the error message if
it's not the case.

Signed-off-by: aszlig <aszlig@nix.build>