flake.nix: Improve syntax linting/type checking
So far, I ran mypy and PHP's linter manually, but to catch errors early,
we really should run this whenever we do "nix flake check", so that we
hopefully avoid pushing bogus code.

I also restructured the Hydra jobs a bit so that the main packages are
also part of the jobset, which makes linting/type errors more obvious
when looking at the evaluation status.

Signed-off-by: aszlig <aszlig@nix.build>