flake.nix: Add and run gdlint on project
While we only had very few code style errors, it does however make sense
to lint our code early on. Of course this is to keep our code clean but
it's also so that we catch errors early and thus prevent us from waiting
until it blows up in our integration test.

This also adds gdtoolkit to our development shell and to the available
packages.

Signed-off-by: aszlig <aszlig@nix.build>