Refactor handling of multiple attachments
So far, file uploads were done by providing an arbitrary key to the
multipart/form-data posted to the proxy, which in turn didn't use the
key name at all but instead wrongly used the *same* key for every file
name.

Of course, fixing this would be easy, by simply not using the same key
and instead re-use the one provided by POST data. However, we also have
a cover_file field, which *referred* to the POST field of the file to be
used as the cover.

This "design" however has a few flaws: If the cover_file points to a
non-existing field (which it did), we'd need to do additional checking
to make sure we actually have a corresponding upload. Additionally, the
fact that we can simply use arbitrary fields, the order in which the
uploads are created is not deterministic.

To fix this, we now no longer accept arbitrary fields but attachments
are now split into "cover" and "attachments".

The first one is optional and might contain a cover image, which also is
validated to make sure it's *really* an image.

The latter accepts one or more attachments of any file type, since it
might also be useful to eg. upload save game state or related data.

In order to make this less cumbersome to use, we can now create
attachments simply by using something like:

  Attachment.from_path("res://some/file.jpg")
  Attachment.from_image(my_shiny_image, "shiny.png")

Right now from_path and from_image are the only constructors, but it's
now pretty easy to just add more ways to construct attachments.

Signed-off-by: aszlig <aszlig@nix.build>