Use PROXY_URL instead of PROXY_{HOST,PATH}
This is mainly to make it more convenient and to have everything at one
place, because for example if you want to test this on a different port,
you need to scroll down into the actual implementation and change the
port accordingly.

The same thing applies if you want to not not use TLS, where you need to
know what the third function argument of HTTPClient.connect_to_host
does.

Unfortunately, Godot doesn't expose an URL parsing function to GDScript,
so we need to implement this by ourselves, which also comes with a bit
of complexity. I do however think it's worth adding this, since it does
make it easier for users to configure.

Signed-off-by: aszlig <aszlig@nix.build>