Remove random identifier in card title
In the original implementation the identifier was used to find the card
after submitting it to Trello. Submitting a card however already returns
the newly created card, so we already know the ID of the card at that
point.

Additionally, we no longer add the identifier to attachments, which
really makes the card's title the only occurence of that identifier and
since it's an unique ID it doesn't really add any value other than being
just random noise.

Signed-off-by: aszlig <aszlig@nix.build>