Encode form values as UTF-8
All the data we submit to Trello is expected to be UTF-8 except the file
attachments, which are not specially encoded anyway.

So all we need to fix this is change .to_ascii() to .to_utf8() for all
the form values.

I also changed main test a little bit to also check whether Unicode
symbols are correctly encoded.

Signed-off-by: aszlig <aszlig@nix.build>