Changed some UI/UX aspects.
The feedback dialogue now hides the form as it is no longer needed and shows a close button.
The button closes the whole dialogue after some feedback is returned (can be successful or not).
I also implemented a function to show the dialogue again and resets all input fields in the "form".

The timer was created as a Node to avoid creating it at runtime.