flake.nix: Fix references to Godot files and proxy
Commit bfc4a4652c180e87fef94b5b828094a835176140 has moved most of the
files into a dedicated "src" directory, breaking the references in the
Nix Flake file.

Since the recommended organisation of a Godot project is that
project.godot resides in the top-level directory, I moved the file back
into it, because resources are located relative to that and we need to
have access to the full project structure from within our tests.

Additionally, I moved back icon.png, since it also doesn't make a lot of
sense to be put in "src", which would indicate it being a source file.

[1]: https://docs.godotengine.org/en/stable/getting_started/workflow/project_setup/project_organization.html

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @RPicster