Fix logo image SHA-256 and size in test
Test failure introduced in 556f7be7b60ceec762193b35c44bda372d7caf27.

The commit in question has changed the logo image and since the SHA-256
and size of the image is verified in the integration test, the test
subsequently fails.

A better way to solve this would be if we'd have a way to properly
parameterise the main functionality instead of expecting users to change
source files. This could probably involve turning the project into a
proper Godot plugin, which I outlined in #2.

For now however, I don't have a lot of time to allocate to dig into
Godot plugins, so I'm simply fixing the SHA-256 and size.

Signed-off-by: aszlig <aszlig@nix.build>